Move WRL converter to tools_misc, update its docs and comments, prefer python3
